% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniSNA-package.R
\docType{package}
\name{aniSNA}
\alias{aniSNA}
\title{aniSNA}
\description{
Obtain network structures from animal GPS telemetry observations and 
statistically analyse them to assess their adequacy for social network analysis. Methods include 
pre-network data permutations, bootstrapping techniques to obtain confidence intervals for global
network metrics, and correlation and regression analysis of the local network metrics.
}
\author{
Prabhleen Kaur
}
