% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elk_2010_permutations}
\alias{elk_2010_permutations}
\title{A list of 100 igraph objects obtained by permuting the raw elk_data_2010 and obtaining network from those}
\format{
A list of 100 igraph objects
}
\usage{
elk_2010_permutations
}
\description{
A list of 100 igraph objects obtained by permuting the raw elk_data_2010 and obtaining network from those
}
\examples{
data(elk_2010_permutations)
}
\keyword{datasets}
