% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elk_all_interactions_2010}
\alias{elk_all_interactions_2010}
\title{Dataset of all possible interactions from elk_data_2010}
\format{
A dataframe with 7615 rows and 5 variables
\describe{
\item{Animal_A}{First animal ID}
\item{Animal_B}{Second animal ID}
\item{Timestamp_A}{Observation timestamp of first animal}
\item{Timestamp_B}{Observation timestamp of second animal}
\item{distance}{Distance in metres between the two animals}
}
}
\usage{
elk_all_interactions_2010
}
\description{
Dataset of all possible interactions from elk_data_2010
}
\examples{
data(elk_all_interactions_2010)
}
\keyword{datasets}
