% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elk_data_2010}
\alias{elk_data_2010}
\title{Data to showcase functions in our package}
\format{
A dataframe with 123568 rows and 4 variables:
\describe{
\item{animal_id}{Unique ID of individuals in the observed sample}
\item{datetime}{Date and timestamp of the observation}
\item{latitude_rad}{Latitude of individual observation in radians}
\item{longitude_rad}{Longitude of individual observation in radians}
}
}
\usage{
elk_data_2010
}
\description{
Contains GPS telemetry observations of the species elk in year 2010
}
\examples{
data(elk_data_2010)
}
\keyword{datasets}
