% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elk_interactions_2010}
\alias{elk_interactions_2010}
\title{Dataset of interactions from elk_data_2010 using first mode as the spatial threshold}
\format{
A dataframe with 2393 rows and 5 variables
\describe{
\item{Animal_A}{First animal ID}
\item{Animal_B}{Second animal ID}
\item{Timestamp_A}{Observation timestamp of first animal}
\item{Timestamp_B}{Observation timestamp of second animal}
\item{distance}{Distance in metres between the two animals}
}
}
\usage{
elk_interactions_2010
}
\description{
Dataset of interactions from elk_data_2010 using first mode as the spatial threshold
}
\examples{
data(elk_interactions_2010)
}
\keyword{datasets}
