% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elk_network_2010}
\alias{elk_network_2010}
\title{An igraph object depicting the network obtained from elk_interactions_2010}
\format{
An igraph object with 57 nodes and 114 edges
}
\usage{
elk_network_2010
}
\description{
An igraph object depicting the network obtained from elk_interactions_2010
}
\examples{
igraph::E(elk_network_2010)
}
\keyword{datasets}
