% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_summary.R
\name{get_network_summary}
\alias{get_network_summary}
\title{Calculates and prints network summary statistics}
\usage{
get_network_summary(network)
}
\arguments{
\item{network}{An undirected network with nodes representing animal IDs and edges representing associations between them.}
}
\description{
Calculates and prints network summary statistics
}
