% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_threshold.R
\name{get_spatial_threshold}
\alias{get_spatial_threshold}
\title{To obtain optimized spatial threshold for calculating interactions from raw GPS observations}
\usage{
get_spatial_threshold(species_interactions)
}
\arguments{
\item{species_interactions}{A dataframe consisting of individual interactions within maximum possible distance}
}
\value{
Spatial threshold in metres
}
\description{
To obtain optimized spatial threshold for calculating interactions from raw GPS observations
}
