% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interacting_pairs}
\alias{interacting_pairs}
\title{Function to obtain pairs of interacting animals}
\usage{
interacting_pairs(
  i,
  datetime,
  latitude,
  longitude,
  temporal_thresh,
  spatial_thresh
)
}
\arguments{
\item{i}{Index of the animal}

\item{datetime}{DateTime vector}

\item{latitude}{latitude vector}

\item{longitude}{longitude vector}

\item{temporal_thresh}{time threshold in minutes}

\item{spatial_thresh}{spatial threshold in meters}
}
\value{
A matrix consisting of two row. The first row corresponds to the interacting indices and the second row to the respective distances.
}
\description{
Function to obtain pairs of interacting animals
}
