% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_from_interactions.R
\name{network_from_interactions}
\alias{network_from_interactions}
\title{Function to obtain a network structure from interactions dataframe}
\usage{
network_from_interactions(species_raw, interactions, n_cores = 1)
}
\arguments{
\item{species_raw}{A dataframe consisting of raw GPS observations}

\item{interactions}{A dataframe of interactions captured through raw GPS observations}

\item{n_cores}{Number of cores for parallel processing, default is 1}
}
\value{
An object of class igraph
}
\description{
Function to obtain a network structure from interactions dataframe
}
\examples{
data(elk_data_2010, elk_interactions_2010)
network_from_interactions(elk_data_2010, elk_interactions_2010)
}
