% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{obtain_confidence_intervals}
\alias{obtain_confidence_intervals}
\title{To obtain 95\% confidence intervals around the observed network statistics}
\usage{
obtain_confidence_intervals(
  network,
  n_versions = 100,
  network_metrics = c("mean_degree", "mean_strength", "density", "diameter",
    "transitivity")
)
}
\arguments{
\item{network}{An igraph object}

\item{n_versions}{Number of bootstrapped versions to be used. (default = 100)}

\item{network_metrics}{Network metrics to be evaluated. This should be supplied as a character vector and the values 
should be chosen from "mean_degree", "mean_strength", "density", "diameter", "transitivity". (default = c("mean_degree", "mean_strength", "density", "diameter", "transitivity"))}
}
\value{
A DataFrame consisting of three columns. The first column contains the value of observed network metric, the second and 
third column represent the lower and upper limit of 95% confidence interval respectively. The rows correspond to the network metrics chosen to evaluate.
}
\description{
To obtain 95\% confidence intervals around the observed network statistics
}
\examples{
\donttest{
data(elk_network_2010)
obtain_confidence_intervals(elk_network_2010, n_versions = 100, 
network_metrics = c("mean_degree", "mean_strength", "density", "diameter", "transitivity"))
}
}
