% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{plot.bootstrapped_pvalue_matrix}
\alias{plot.bootstrapped_pvalue_matrix}
\title{To plot the results obtained from obtain_two_bs_versions function}
\usage{
\method{plot}{bootstrapped_pvalue_matrix}(x, ...)
}
\arguments{
\item{x}{A matrix of p-values obtained from obtain_two_bs_versions function}

\item{...}{Further arguments are ignored.}
}
\description{
To plot the results obtained from obtain_two_bs_versions function
}
\examples{
data(elk_network_2010)
mean_pvalue_matrix <- obtain_two_bs_versions(elk_network_2010, n_versions = 100)
plot(mean_pvalue_matrix)
}
