% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_level_confidence_intervals.R
\name{plot.list_node_level_CI}
\alias{plot.list_node_level_CI}
\title{To plot the results for node-level confidence intervals}
\usage{
\method{plot}{list_node_level_CI}(x, ...)
}
\arguments{
\item{x}{A list of dataframes obtained from obtain_node_level_CI function.}

\item{...}{Further arguments are ignored.}
}
\value{
No return value, called for side effects. 
The plots show 95% confidence intervals along with the observed metric value for each of the nodes in the network.
}
\description{
To plot the results for node-level confidence intervals
}
\examples{
\donttest{
data(elk_network_2010)
elk_node_level_CI <- obtain_node_level_CI(elk_network_2010)
plot(elk_node_level_CI)
}
}
