% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{plot_correlation_results}
\alias{plot_correlation_results}
\title{To plot correlation analysis results}
\usage{
plot_correlation_results(correlation_results)
}
\arguments{
\item{correlation_results}{A list of matrices obtained from correlation_analyze function}
}
\description{
To plot correlation analysis results
}
