% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_summary.R
\name{plot_network}
\alias{plot_network}
\title{Visualize Animal Network}
\usage{
plot_network(species_network, seed = 1)
}
\arguments{
\item{species_network}{An igraph graph object consisting of observed network.}

\item{seed}{Seed to be set for layout.}
}
\value{
No return value, called for side effects. The plots depict a visualisation of network structure.
}
\description{
Visualize Animal Network
}
\examples{
data(elk_network_2010)
plot_network(elk_network_2010)
}
