% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{plot_p_values_bootstrapped_versions}
\alias{plot_p_values_bootstrapped_versions}
\title{To plot the results obtained from obtain_two_bs_versions function}
\usage{
plot_p_values_bootstrapped_versions(bootstrapped_results)
}
\arguments{
\item{bootstrapped_results}{A matrix of p-values obtained from obtain_two_bs_versions function}
}
\description{
To plot the results obtained from obtain_two_bs_versions function
}
