% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_analysis.R
\name{plot_regression_results}
\alias{plot_regression_results}
\title{To plot regression analysis results}
\usage{
plot_regression_results(regression_results)
}
\arguments{
\item{regression_results}{A list of matrices obtained from regression_slope_analyze function}
}
\description{
To plot regression analysis results
}
