% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_analysis.R
\name{plot_subsamples}
\alias{plot_subsamples}
\title{To plot sub-sampling results}
\usage{
plot_subsamples(result, network)
}
\arguments{
\item{result}{A list of matrices obtained from network_subsamples function}

\item{network}{An igraph graph object consisting of observed network}
}
\description{
To plot sub-sampling results
}
