% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{plot_width_CI}
\alias{plot_width_CI}
\title{Title To plot the results obtained from width_CI function}
\usage{
plot_width_CI(width_CI_results)
}
\arguments{
\item{width_CI_results}{A matrix of width of Confidence Intervals obtained from width_CI function}
}
\description{
Title To plot the results obtained from width_CI function
}
