% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{width_CI}
\alias{width_CI}
\title{To obtain width of confidence intervals using bootstrapped versions at each level of sub-sampling}
\usage{
width_CI(
  network,
  n_versions = 1000,
  seed = 12345,
  n.iter = 10,
  network_metrics = c("mean_degree", "mean_strength", "density", "diameter",
    "transitivity")
)
}
\arguments{
\item{network}{An igraph object}

\item{n_versions}{Number of bootstrapped versions to be used}

\item{seed}{seed number}

\item{n.iter}{Number of iterations at each level}

\item{network_metrics}{Network metrics to be evaluated. This should be supplied as a character vector and the values 
should be chosen from "mean_degree", "mean_strength", "density", "diameter", "transitivity". (default = c("mean_degree", "mean_strength", "density", "diameter", "transitivity")).}
}
\value{
A matrix containing width of Confidence Intervals where each row corresponds to the sub-sample size and columns correspond to the chosen network metric.
}
\description{
To obtain width of confidence intervals using bootstrapped versions at each level of sub-sampling
}
