% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_radian_coordinates}
\alias{distance_radian_coordinates}
\title{Calculate distance between two pairs of radian coordinates}
\usage{
distance_radian_coordinates(latf, lonf, latt, lont)
}
\arguments{
\item{latf}{latitude from}

\item{lonf}{longitude from}

\item{latt}{latitude to}

\item{lont}{longitude to}
}
\value{
distance value in meters
}
\description{
Calculate distance between two pairs of radian coordinates
}
