% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_summary.R
\name{get_network_summary}
\alias{get_network_summary}
\title{Calculates and prints network summary statistics}
\usage{
get_network_summary(network)
}
\arguments{
\item{network}{An undirected network with nodes representing animal IDs and edges representing associations between them.}
}
\value{
No return value, called for side effects. The function prints values of network metrics to the console.
}
\description{
Calculates and prints network summary statistics
}
\examples{
data(elk_network_2010)
get_network_summary(elk_network_2010)
}
