% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{plot.bootstrapped_pvalue_matrix}
\alias{plot.bootstrapped_pvalue_matrix}
\title{To plot the results obtained from bootstrapped_difference_pvalues function}
\usage{
\method{plot}{bootstrapped_pvalue_matrix}(x, ...)
}
\arguments{
\item{x}{A matrix of p-values obtained from bootstrapped_difference_pvalues function}

\item{...}{Further arguments are ignored.}
}
\value{
No return value, called for side effects. The plot shows p-values between 0 and 1 corresponding to each sample size.
}
\description{
To plot the results obtained from bootstrapped_difference_pvalues function
}
\examples{
\donttest{
data(elk_network_2010)
mean_pvalue_matrix <- bootstrapped_difference_pvalues(elk_network_2010, n_versions = 100)
plot(mean_pvalue_matrix)
}
}
