% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{plot.list_correlation_matrices}
\alias{plot.list_correlation_matrices}
\title{To plot correlation analysis results}
\usage{
\method{plot}{list_correlation_matrices}(x, ...)
}
\arguments{
\item{x}{A list of matrices obtained from correlation_analyze function}

\item{...}{Further arguments are ignored}
}
\description{
To plot correlation analysis results
}
