% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_analysis.R
\name{plot.list_regression_matrices}
\alias{plot.list_regression_matrices}
\title{To plot regression analysis results}
\usage{
\method{plot}{list_regression_matrices}(x, ...)
}
\arguments{
\item{x}{A list of matrices obtained from regression_slope_analyze function}

\item{...}{Further arguments are ignored}
}
\value{
No return value, called for side effects. The plots show regression slope values corresponding to proportion of individuals in the sample.
}
\description{
To plot regression analysis results
}
\examples{
\donttest{
data(elk_network_2010)
elk_regression_analysis <- regression_slope_analyze(elk_network_2010)
plot(elk_regression_analysis)
}
}
