% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_analysis.R
\name{plot.Subsampled_Network_Metrics}
\alias{plot.Subsampled_Network_Metrics}
\title{To plot sub-sampling results}
\usage{
\method{plot}{Subsampled_Network_Metrics}(x, network, ...)
}
\arguments{
\item{x}{A list of matrices obtained from network_subsamples function of class "Subsampled_Network_Metrics"}

\item{network}{An igraph graph object consisting of observed network}

\item{...}{Further arguments are ignored}
}
\value{
No return value, called for side effects. The boxplots depict range of values, network metrics take when multiple subsamples are chosen from the observed sample.
}
\description{
To plot sub-sampling results
}
\examples{

data(elk_network_2010)
elk_subsamples <- network_subsamples(elk_network_2010)
plot(elk_subsamples, elk_network_2010)

}
