% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_permuted_networks.R
\name{plot.Subsampled_Permuted_Network_Metrics}
\alias{plot.Subsampled_Permuted_Network_Metrics}
\title{To plot sub-sampling results of the original network and permuted networks}
\usage{
\method{plot}{Subsampled_Permuted_Network_Metrics}(x, network, ...)
}
\arguments{
\item{x}{A list of matrices obtained from subsamples_permuted_networks function of class "Subsampled_Permuted_Network_Metrics"}

\item{network}{An igraph graph object consisting of observed network}

\item{...}{Further arguments are ignored}
}
\value{
No return value, called for side effects. The boxplots show side-by-side comparison of network metrics distribution from subsamples of observed network and subsamples from permuted networks.
}
\description{
To plot sub-sampling results of the original network and permuted networks
}
\examples{
\donttest{
data(elk_2010_permutations, elk_network_2010)
elk_subsamples_permuted_networks <- subsamples_permuted_networks(elk_2010_permutations)
plot(elk_subsamples_permuted_networks, elk_network_2010)
}
}
