% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{plot.Width_CI_matrix}
\alias{plot.Width_CI_matrix}
\title{To plot the results obtained from width_CI function}
\usage{
\method{plot}{Width_CI_matrix}(x, ...)
}
\arguments{
\item{x}{A matrix of width of Confidence Intervals obtained from width_CI function}

\item{...}{Further arguments are ignored.}
}
\value{
No return value, called for side effects. Plots show width of confidence intervals corresponding to number of individuals in the sub-sample.
}
\description{
To plot the results obtained from width_CI function
}
\examples{
\donttest{
data(elk_network_2010)
width_CI_elk <- width_CI(elk_network_2010, n_versions = 100)
plot(width_CI_elk)
}
}
